Package renaming and folder restructuring
Rename repository to infrastructure and move google map client files under infrastructure folder
