+++ httpserver/router.go
@@ -2,9 +2,9 @@ package httpserver
 
 import (
 	"github.com/imylam/delivery-test/db"
-	"github.com/imylam/delivery-test/googlemap"
 	_orderHandler "github.com/imylam/delivery-test/order/api/rest"
-	_orderRepo "github.com/imylam/delivery-test/order/repository/mysql"
+	"github.com/imylam/delivery-test/order/infrastructure/googlemap"
+	_orderRepo "github.com/imylam/delivery-test/order/infrastructure/mysql"
 	_orderUsecase "github.com/imylam/delivery-test/order/usecase"
 
 	"github.com/gin-gonic/gin"
